Make Admission own identity component decisions
